requirements.txt
Flask
scikit-learn
Pillow
python-dotenv
requests
numpy
werkzeug
imbalanced-learn
gunicorn
Flask-Cors